Implement IO device battery sensor as binary sensor
